+++ lib/dartapi_core.dart
@@ -6,5 +6,6 @@ export 'src/core/api_route.dart';
 export 'src/core/base_controller.dart';
 export 'src/core/logger.dart';
 export 'src/core/serializable.dart';
+export 'src/core/global_exception_handler.dart';
 
 export 'src/utils/utils.dart';